pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "study-buddy"
version = "0.1.0"
description = "Agentic pipeline that converts study materials into flashcards, decks, and cheat sheets."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Study Buddy Maintainers" }]
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "jinja2>=3.1,<4.0",
  "python-multipart>=0.0.9,<0.1",
  "pydantic>=2.9,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "python-dotenv>=1.0,<2.0",
  "langchain-core>=0.3.15,<0.4",
  "langchain-openai>=0.2.14,<0.3",
  "langchain-text-splitters>=0.3.2,<0.4",
  "langchain-chroma>=0.1.4,<0.2",
  "langchain-community>=0.3.13,<0.4",
  "chromadb>=0.5.3,<0.6",
  "crewai>=0.193.2,<0.194",
  "pypdf>=4.3,<5.0",
  "markdownify>=1.2,<2.0",
  "python-pptx>=0.6.23,<0.7",
  "mammoth>=1.11,<2.0",
  "genanki>=0.13,<0.14",
  "firecrawl>=4.3,<5.0",
  "tavily-python>=0.7.12,<0.8",
]

[project.optional-dependencies]
test = ["pytest>=8.3,<9.0", "pytest-asyncio>=0.25,<0.26"]
dev = ["ruff>=0.7,<0.8"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["venv", "chroma_db", "uploads", "outputs"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E203", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
